+++ app/build.gradle
@@ -57,7 +57,7 @@ dependencies {
     implementation 'androidx.test.espresso:espresso-idling-resource:3.2.0'
     testImplementation 'junit:junit:4.13'
     testImplementation 'org.mockito:mockito-core:3.3.0'
-    testCompile "org.mockito:mockito-inline:3.3.0"
+    testImplementation "org.mockito:mockito-inline:3.3.0"
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
     androidTestImplementation 'androidx.test.espresso:espresso-web:3.2.0'
