--- app/src/main/java/ch/epfl/sdp/Firestore.java
@@ -1,6 +1,0 @@
-package ch.epfl.sdp;
-
-public interface Firestore {
-    void readDocument(Callback callback);
-    void writeDocument(Callback callback);
-}
