+++ app/src/main/java/ch/epfl/sdp/FirestoreWrapper.java
@@ -4,6 +4,8 @@ import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.firestore.DocumentReference;
+import com.google.firebase.firestore.DocumentSnapshot;
+import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.QuerySnapshot;
 
 import java.io.Serializable;
@@ -11,7 +13,7 @@ import java.util.Map;
 
 public interface FirestoreWrapper extends Serializable {
 
-    public <A,B> FirestoreWrapper add(Map<A, B> map);
+    public <A, B> FirestoreWrapper add(Map<A, B> map);
 
     public FirestoreWrapper collection(String collectionPath);
